requirements.txt
data==0.4
numpy==1.19.5
Pillow==9.5.0
torch==1.8.0+cu111
torchvision==0.9.0+cu111
blobfile==0.11.0
mpi4py==3.1.4
tensorflow==1.14.0
tqdm==4.64.1
wandb==0.13.10
scipy==1.5.4
